# backend_repo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/trainer": {
            "post": {
                "description": "Позволяет добавить нового тренера в систему",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trainers"
                ],
                "summary": "Добавление нового тренера",
                "parameters": [
                    {
                        "description": "Данные тренера",
                        "name": "trainer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trainer_handlers.trainerReqHandler"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренер успешно создан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ввод",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании тренера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/trainer": {
            "put": {
                "description": "Позволяет обновить информацию о тренере",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trainers"
                ],
                "summary": "Обновляет данные тренера",
                "parameters": [
                    {
                        "description": "Информация о тренере для обновления",
                        "name": "trainer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trainer_handlers.trainerUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновление выполнено успешно",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ввод",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера при обновлении",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет тренера по его идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trainers"
                ],
                "summary": "Удаление тренера",
                "parameters": [
                    {
                        "description": "Идентификатор тренера для удаления",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trainer_handlers.deleteHandler"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Тренер успешно удален",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ввод",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Тренер не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера при удалении тренера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "description": "Retrieves a single user by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_handler.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Schedule": {
            "type": "object",
            "required": [
                "description",
                "max_participants",
                "min_participants",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEnable": {
                    "type": "boolean"
                },
                "max_participants": {
                    "type": "integer"
                },
                "min_participants": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "trainers_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "users_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Subscription": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "trainer_handlers.deleteHandler": {
            "type": "object"
        },
...
        "trainer_handlers.trainerReqHandler": {
            "type": "object"
        },
        "trainer_handlers.trainerUpdateReq": {
            "type": "object"
        },
        "user_handler.UserResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passport": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Schedule"
                    }
                },
                "status": {
                    "type": "string"
                },
                "subscription": {
                    "$ref": "#/definitions/models.Subscription"
                }
            }
        }
    }
}